Removed coverartdelegate and renamed basecoverartdelegate to coverartdelegate
