+++ src/library/basetracktablemodel.h
@@ -9,7 +9,7 @@
 #include "library/trackmodel.h"
 #include "track/track_decl.h"
 
-class BaseCoverArtDelegate;
+class CoverArtDelegate;
 class TrackCollectionManager;
 
 class BaseTrackTableModel : public QAbstractTableModel, public TrackModel {
@@ -188,7 +188,7 @@ class BaseTrackTableModel : public QAbstractTableModel, public TrackModel {
     // the internal database.
     virtual TrackId doGetTrackId(
             const TrackPointer& pTrack) const;
-    virtual BaseCoverArtDelegate* doCreateCoverArtDelegate(
+    virtual CoverArtDelegate* doCreateCoverArtDelegate(
             QTableView* pTableView) const = 0;
 
     QVariant composeCoverArtToolTipHtml(
